Make getStoreTxFromDidDocument only used in tests
